[DEV-10253] Moving more integration tests to the right place and adding check in travis if they are improperly categorized
